requirements.txt
fire
tqdm
jsonlines
path
requests
gdown
feedparser
html2text
markdownify
GitPython
pypandoc
epub_meta
arxiv
pandas
grobid_tei_xml
python-docx
transformers
openpyxl
seaborn
matplotlib
codaio
python-dotenv
PyPDF2
PyCryptodome
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
